+++ util/opentelemetry-util-genai/CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Add jsonlines support to fsspec uploader
+  ([https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3791](#3791))
+- Rename "fsspec_upload" entry point and classes to more generic "upload"
+  ([https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3798](#3798))
+- Record content-type and use canonical paths in fsspec genai uploader
+  ([https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3795](#3795))
+
 ## Version 0.1b0 (2025-09-25)
 
 - Add completion hook to genai utils to implement semconv v1.37.